refactor: SettingsViewModel -> ThemeViewModel로 수정
