+++ lib/features/chart/views/widgets/mood_frequency_card.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:moodtracker/core/widgets/center_progress_indicator.dart';
 import 'package:moodtracker/core/widgets/center_text.dart';
-import 'package:moodtracker/features/chart/provider/provider.dart';
+import 'package:moodtracker/features/chart/providers/provider.dart';
 import 'package:moodtracker/features/chart/views/widgets/mood_frequency_card_item.dart';
 
 class MoodFrequencyCard extends ConsumerWidget {
