+++ lib/features/settings/view_models/theme_view_model.dart
@@ -1,14 +1,10 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:moodtracker/core/di/provider.dart';
-import 'package:moodtracker/core/repositories/interfaces/theme_repository.dart';
 import 'package:moodtracker/core/theme/app_theme_type.dart';
 
-class SettingsViewModel {
-  final ThemeRepository themeRepo;
-
-  SettingsViewModel({required this.themeRepo});
-
-  AsyncValue<AppThemeType> getTheme(WidgetRef ref) {
+class ThemeViewModel extends AutoDisposeNotifier<AsyncValue<AppThemeType>> {
+  @override
+  AsyncValue<AppThemeType> build() {
     return ref.watch(themeRepository).when(
           data: (data) {
             return AsyncValue.data(data.theme);
@@ -19,12 +15,6 @@ class SettingsViewModel {
   }
 
   void selectTheme(AppThemeType theme) {
-    themeRepo.setTheme(theme);
+    ref.read(themeRepository.notifier).setTheme(theme);
   }
 }
-
-final settingsProvider = Provider<SettingsViewModel>(
-  (ref) => SettingsViewModel(
-    themeRepo: ref.watch(themeRepository.notifier),
-  ),
-);
