+++ lib/features/settings/views/settings_screen.dart
@@ -5,7 +5,7 @@ import 'package:go_router/go_router.dart';
 import 'package:moodtracker/core/theme/app_theme_type.dart';
 import 'package:moodtracker/core/widgets/center_progress_indicator.dart';
 import 'package:moodtracker/core/widgets/dialog/error_dialog.dart';
-import 'package:moodtracker/features/settings/view_models/settings_view_model.dart';
+import 'package:moodtracker/features/settings/providers/provider.dart';
 import 'package:moodtracker/features/settings/views/widgets/settings_item.dart';
 import 'package:moodtracker/features/settings/views/widgets/theme_menu_item.dart';
 import 'package:moodtracker/route/route_path.dart';
@@ -19,7 +19,6 @@ class SettingsScreen extends ConsumerStatefulWidget {
 
 class _SettingsScreenState extends ConsumerState<SettingsScreen> {
   final MenuController _menuController = MenuController();
-  late final SettingsViewModel _viewModel = ref.read(settingsProvider);
 
   @override
   Widget build(BuildContext context) {
@@ -29,7 +28,7 @@ class _SettingsScreenState extends ConsumerState<SettingsScreen> {
         SettingsItem(
           onTap: () {},
           text: "Theme".tr(),
-          trailing: _viewModel.getTheme(ref).when(
+          trailing: ref.watch(themeProvider).when(
             data: (currentTheme) {
               return MenuAnchor(
                 controller: _menuController,
@@ -82,7 +81,7 @@ class _SettingsScreenState extends ConsumerState<SettingsScreen> {
 
   void _selectTheme(AppThemeType theme) {
     try {
-      _viewModel.selectTheme(theme);
+      ref.read(themeProvider.notifier).selectTheme(theme);
     } catch (e) {
       showErrorDialog(
         context: context,
